Added prioritization level 1 and 2 quantities for all five imaging modes of the IKZ Apreo, tested generation of NeXus files for all five test images from Robert Kernke successfully, next step continue with eds
