+++ imgs.batch.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-datasource="../../../../paper_paper_paper/scidat_nomad_ebsd/bb_analysis/data/development_imgs/"
+datasource="../../../../paper_paper_paper/scidat_nomad_ebsd/bb_analysis/data/development_imgs/ikz_robert/"
 
 # comments is detector mode
 examples="kit/FeMoOx_AntiA_04_1k5x_CN.tif"
@@ -9,6 +9,8 @@ examples="ikz_martin/ALN_baoh_021.tif"  # T2
 examples="ikz_robert/T3_image.tif"
 examples="ikz_robert/ETD_image.tif"  # ETD
 examples="ikz_martin/NavCam_normal_vis_light_ccd.tif"  # NavCam
+
+examples="0c8nA_3deg_003_AplusB_test.tif ALN_baoh_021.tif T3_image.tif ETD_image.tif NavCam_normal_vis_light_ccd.tif"
 
 
 for example in $examples; do
