+++ pynxtools/dataconverter/readers/em/reader.py
@@ -22,7 +22,7 @@
 from typing import Tuple, Any
 
 from pynxtools.dataconverter.readers.base.reader import BaseReader
-from pynxtools.pynxtools.dataconverter.readers.em.concepts.nxs_concepts import NxEmAppDef
+from pynxtools.dataconverter.readers.em.concepts.nxs_concepts import NxEmAppDef
 # from pynxtools.dataconverter.readers.em.subparsers.nxs_mtex import NxEmNxsMTexSubParser
 # from pynxtools.dataconverter.readers.em.subparsers.nxs_pyxem import NxEmNxsPyxemSubParser
 from pynxtools.dataconverter.readers.em.subparsers.nxs_imgs import NxEmImagesSubParser
