+++ pynxtools/dataconverter/readers/em/subparsers/image_tiff_tfs_cfg.py
@@ -17,397 +17,35 @@
 #
 """Configuration of the image_tiff_tfs subparser."""
 
-from typing import List
-
-
-TiffTfsConcepts = ["User/Date",
-                   "User/Time",
-                   "User/User",
-                   "User/UserText",
-                   "User/UserTextUnicode",
-                   "System/Type",
-                   "System/Dnumber",
-                   "System/Software",
-                   "System/BuildNr",
-                   "System/Source",
-                   "System/Column",
-                   "System/FinalLens",
-                   "System/Chamber",
-                   "System/Stage",
-                   "System/Pump",
-                   "System/ESEM",
-                   "System/Aperture",
-                   "System/Scan",
-                   "System/Acq",
-                   "System/EucWD",
-                   "System/SystemType",
-                   "System/DisplayWidth",
-                   "System/DisplayHeight",
-                   "Beam/HV",
-                   "Beam/Spot",
-                   "Beam/StigmatorX",
-                   "Beam/StigmatorY",
-                   "Beam/BeamShiftX",
-                   "Beam/BeamShiftY",
-                   "Beam/ScanRotation",
-                   "Beam/ImageMode",
-                   "Beam/FineStageBias",
-                   "Beam/Beam",
-                   "Beam/Scan",
-                   "EBeam/Source",
-                   "EBeam/ColumnType",
-                   "EBeam/FinalLens",
-                   "EBeam/Acq",
-                   "EBeam/Aperture",
-                   "EBeam/ApertureDiameter",
-                   "EBeam/HV",
-                   "EBeam/HFW",
-                   "EBeam/VFW",
-                   "EBeam/WD",
-                   "EBeam/BeamCurrent",
-                   "EBeam/TiltCorrectionIsOn",
-                   "EBeam/DynamicFocusIsOn",
-                   "EBeam/DynamicWDIsOn",
-                   "EBeam/ScanRotation",
-                   "EBeam/LensMode",
-                   "EBeam/LensModeA",
-                   "EBeam/ATubeVoltage",
-                   "EBeam/UseCase",
-                   "EBeam/SemOpticalMode",
-                   "EBeam/ImageMode",
-                   "EBeam/SourceTiltX",
-                   "EBeam/SourceTiltY",
-                   "EBeam/StageX",
-                   "EBeam/StageY",
-                   "EBeam/StageZ",
-                   "EBeam/StageR",
-                   "EBeam/StageTa",
-                   "EBeam/StageTb",
-                   "EBeam/StigmatorX",
-                   "EBeam/StigmatorY",
-                   "EBeam/BeamShiftX",
-                   "EBeam/BeamShiftY",
-                   "EBeam/EucWD",
-                   "EBeam/EmissionCurrent",
-                   "EBeam/TiltCorrectionAngle",
-                   "EBeam/PreTilt",
-                   "EBeam/WehneltBias",
-                   "EBeam/BeamMode",
-                   "EBeam/MagnificationCorrection",
-                   "GIS/Number",
-                   "Scan/InternalScan",
-                   "Scan/Dwelltime",
-                   "Scan/PixelWidth",
-                   "Scan/PixelHeight",
-                   "Scan/HorFieldsize",
-                   "Scan/VerFieldsize",
-                   "Scan/Average",
-                   "Scan/Integrate",
-                   "Scan/FrameTime",
-                   "EScan/Scan",
-                   "EScan/InternalScan",
-                   "EScan/Dwell",
-                   "EScan/PixelWidth",
-                   "EScan/PixelHeight",
-                   "EScan/HorFieldsize",
-                   "EScan/VerFieldsize",
-                   "EScan/FrameTime",
-                   "EScan/LineTime",
-                   "EScan/Mainslock",
-                   "EScan/LineIntegration",
-                   "EScan/ScanInterlacing",
-                   "Stage/StageX",
-                   "Stage/StageY",
-                   "Stage/StageZ",
-                   "Stage/StageR",
-                   "Stage/StageT",
-                   "Stage/StageTb",
-                   "Stage/SpecTilt",
-                   "Stage/WorkingDistance",
-                   "Stage/ActiveStage",
-                   "Image/DigitalContrast",
-                   "Image/DigitalBrightness",
-                   "Image/DigitalGamma",
-                   "Image/Average",
-                   "Image/Integrate",
-                   "Image/ResolutionX",
-                   "Image/ResolutionY",
-                   "Image/DriftCorrected",
-                   "Image/ZoomFactor",
-                   "Image/ZoomPanX",
-                   "Image/ZoomPanY",
-                   "Image/MagCanvasRealWidth",
-                   "Image/MagnificationMode",
-                   "Image/ScreenMagCanvasRealWidth",
-                   "Image/ScreenMagnificationMode",
-                   "Image/PostProcessing",
-                   "Image/Transformation",
-                   "Vacuum/ChPressure",
-                   "Vacuum/Gas",
-                   "Vacuum/UserMode",
-                   "Vacuum/Humidity",
-                   "Specimen/Temperature",
-                   "Specimen/SpecimenCurrent",
-                   "Detectors/Number",
-                   "Detectors/Name",
-                   "Detectors/Mode",
-                   "T2/Contrast",
-                   "T2/Brightness",
-                   "T2/Signal",
-                   "T2/ContrastDB",
-                   "T2/BrightnessDB",
-                   "T2/Setting",
-                   "T2/MinimumDwellTime",
-                   "Accessories/Number",
-                   "EBeamDeceleration/ModeOn",
-                   "EBeamDeceleration/LandingEnergy",
-                   "EBeamDeceleration/ImmersionRatio",
-                   "EBeamDeceleration/StageBias",
-                   "CompoundLensFilter/IsOn",
-                   "CompoundLensFilter/ThresholdEnergy",
-                   "PrivateFei/BitShift",
-                   "PrivateFei/DataBarSelected",
-                   "PrivateFei/DataBarAvailable",
-                   "PrivateFei/TimeOfCreation",
-                   "PrivateFei/DatabarHeight",
-                   "HiResIllumination/BrightFieldIsOn",
-                   "HiResIllumination/BrightFieldValue",
-                   "HiResIllumination/DarkFieldIsOn",
-                   "HiResIllumination/DarkFieldValue",
-                   "EasyLift/Rotation",
-                   "HotStageMEMS/HeatingCurrent",
-                   "HotStageMEMS/HeatingVoltage",
-                   "HotStageMEMS/TargetTemperature",
-                   "HotStageMEMS/ActualTemperature",
-                   "HotStageMEMS/HeatingPower",
-                   "HotStageMEMS/SampleBias",
-                   "HotStageMEMS/SampleResistance",
-                   "HotStage/TargetTemperature",
-                   "HotStage/ActualTemperature",
-                   "HotStage/SampleBias",
-                   "HotStage/ShieldBias",
-                   "HotStageHVHS/TargetTemperature",
-                   "HotStageHVHS/ActualTemperature",
-                   "HotStageHVHS/SampleBias",
-                   "HotStageHVHS/ShieldBias",
-                   "ColdStage/TargetTemperature",
-                   "ColdStage/ActualTemperature",
-                   "ColdStage/Humidity",
-                   "ColdStage/SampleBias"]
-
-# there is more to know and understand than just knowing TFS/FEI uses
-# the above-mentioned concepts in their taxonomy:
-# take the example of System/Source for which an example file (instance) has the
-# value "FEG"
-# similar like in NeXus "System/Source" labels a concept for which (assumption!) there
-# is a controlled enumeration of symbols possible (as the example shows "FEG" is one such
-# allowed symbol of the enumeration.
-# The key issue is that the symbols for the leaf (here "FEG") means nothing eventually
-# when one has another semantic world-view, like in NOMAD metainfo or NeXus
-# (only us) humans understand that what TFS/FEI likely means with the symbol
-# "FEG" is exactly the same as what we mean in NeXus when setting emitter_type of
-# NXebeam_column to "cold_cathode_field_emitter"
-# world with the controlled enumeration value "other" because we do not know
-# if FEG means really a filament or a cold_cathode_field_emitter
 
 TfsToNexusConceptMapping = {"System/Source/FEG": "cold_field_cathode_emitter"}
 
 
-def get_fei_parent_concepts() -> List:
-    """Get list of unique FEI parent concepts."""
-    parent_concepts = set()
-    for entry in TiffTfsConcepts:
-        if isinstance(entry, str) and entry.count("/") == 1:
-            parent_concepts.add(entry.split("/")[0])
-    return list(parent_concepts)
-
-
-def get_fei_childs(concept: str) -> List:
-    """Get all children of FEI parent concept."""
-    child_concepts = set()
-    for entry in TiffTfsConcepts:
-        if isinstance(entry, str) and entry.count("/") == 1:
-            if entry.startswith(f"{concept}/") is True:
-                child_concepts.add(entry.split("/")[1])
-    return list(child_concepts)
-
-
 # "/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/start_time"
-TiffTfsToNeXusCfg = {"IGNORE": {"fun": "ikz_berlin_apreo_iso8601", "terms": ["User/Date", "User/Time"]},
-                     "IGNORE": {"fun": "load_from", "terms": "User/User"},
-                     "IGNORE": {"fun": "load_from", "terms": "User/UserText"},
-                     "IGNORE": {"fun": "load_from", "terms": "User/UserTextUnicode"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Type"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Dnumber"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Software"},
-                     "/ENTRY[entry*]/measurement/em_lab/FABRICATION[fabrication]/identifier": {"fun": "load_from", "terms": "System/BuildNr"},
-                     "/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/electron_source/emitter_type": {"fun": "tfs_to_nexus", "terms": "System/Source"},
-                     "/ENTRY[entry*]/measurement/em_lab/FABRICATION[fabrication]/vendor": "FEI",
-                     "IGNORE": {"fun": "load_from", "terms": "System/FinalLens"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Chamber"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Stage"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Pump"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/ESEM"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Aperture"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Scan"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/Acq"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/EucWD"},
-                     "/ENTRY[entry*]/measurement/em_lab/FABRICATION[fabrication]/model": {"fun": "load_from", "terms": "System/SystemType"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/DisplayWidth"},
-                     "IGNORE": {"fun": "load_from", "terms": "System/DisplayHeight"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/HV"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/Spot"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/StigmatorX"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/StigmatorY"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/BeamShiftX"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/BeamShiftY"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/ScanRotation"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/ImageMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/FineStageBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/Beam"},
-                     "IGNORE": {"fun": "load_from", "terms": "Beam/Scan"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/Source"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/ColumnType"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/FinalLens"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/Acq"},
-                     "/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/APERTURE_EM[aperture_em*]/description": {"fun": "load_from", "terms": "EBeam/Aperture"},
-                     "/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/APERTURE_EM[aperture_em*]/value": {"fun": "load_from", "terms": "EBeam/ApertureDiameter"},
-                     "/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/APERTURE_EM[aperture_em*]/value/@units": "m",
-                     "/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/electron_source/voltage": {"fun": "load_from", "terms": "EBeam/HV"},
-                     "/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/electron_source/voltage/@units": "V",
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/HFW"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/VFW"},
-                     "/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/working_distance": {"fun": "load_from", "terms": "EBeam/WD"},
-                     "/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/working_distance/@units": "m",
-                     "/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/beam_current": {"fun": "load_from", "terms": "EBeam/WD"},
-                     "/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/beam_current/@units": "A",
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/TiltCorrectionIsOn"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/DynamicFocusIsOn"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/DynamicWDIsOn"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/ScanRotation"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/LensMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/LensModeA"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/ATubeVoltage"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/UseCase"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/SemOpticalMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/ImageMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/SourceTiltX"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/SourceTiltY"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StageX"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StageY"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StageZ"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StageR"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StageTa"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StageTb"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StigmatorX"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/StigmatorY"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/BeamShiftX"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/BeamShiftY"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/EucWD"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/EmissionCurrent"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/TiltCorrectionAngle"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/PreTilt"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/WehneltBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/BeamMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeam/MagnificationCorrection"},
-                     "IGNORE": {"fun": "load_from", "terms": "GIS/Number"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/InternalScan"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/Dwelltime"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/PixelWidth"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/PixelHeight"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/HorFieldsize"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/VerFieldsize"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/Average"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/Integrate"},
-                     "IGNORE": {"fun": "load_from", "terms": "Scan/FrameTime"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/Scan"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/InternalScan"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/Dwell"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/PixelWidth"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/PixelHeight"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/HorFieldsize"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/VerFieldsize"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/FrameTime"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/LineTime"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/Mainslock"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/LineIntegration"},
-                     "IGNORE": {"fun": "load_from", "terms": "EScan/ScanInterlacing"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/StageX"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/StageY"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/StageZ"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/StageR"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/StageT"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/StageTb"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/SpecTilt"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/WorkingDistance"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ActiveStage"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/[Image]"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/DigitalContrast"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/DigitalBrightness"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/DigitalGamma"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/Average"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/Integrate"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ResolutionX"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ResolutionY"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/DriftCorrected"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ZoomFactor"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ZoomPanX"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ZoomPanY"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/MagCanvasRealWidth"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/MagnificationMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ScreenMagCanvasRealWidth"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/ScreenMagnificationMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/PostProcessing"},
-                     "IGNORE": {"fun": "load_from", "terms": "StageX/Transformation"},
-                     "IGNORE": {"fun": "load_from", "terms": "Vacuum/ChPressure"},
-                     "IGNORE": {"fun": "load_from", "terms": "Vacuum/Gas"},
-                     "IGNORE": {"fun": "load_from", "terms": "Vacuum/UserMode"},
-                     "IGNORE": {"fun": "load_from", "terms": "Vacuum/Humidity"},
-                     "IGNORE": {"fun": "load_from", "terms": "Specimen/Temperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "Specimen/SpecimenCurrent"},
-                     "IGNORE": {"fun": "load_from", "terms": "Detectors/Number"},
-                     "/ENTRY[entry*]/measurement/em_lab/DETECTOR[detector*]/local_name": {"fun": "load_from", "terms": "Detectors/Name"},
-                     "IGNORE": {"fun": "load_from", "terms": "Detectors/Mode"},
-                     "IGNORE": {"fun": "load_from", "terms": "T2/Contrast"},
-                     "IGNORE": {"fun": "load_from", "terms": "T2/Brightness"},
-                     "/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/event_type": {"fun": "load_from_lower_case", "terms": "T2/Signal"},
-                     "IGNORE": {"fun": "load_from", "terms": "T2/ContrastDB"},
-                     "IGNORE": {"fun": "load_from", "terms": "T2/BrightnessDB"},
-                     "IGNORE": {"fun": "load_from", "terms": "T2/Setting"},
-                     "IGNORE": {"fun": "load_from", "terms": "T2/MinimumDwellTime"},
-                     "IGNORE": {"fun": "load_from", "terms": "Accessories/Number"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeamDeceleration/ModeOn"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeamDeceleration/LandingEnergy"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeamDeceleration/ImmersionRatio"},
-                     "IGNORE": {"fun": "load_from", "terms": "EBeamDeceleration/StageBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "CompoundLensFilter/IsOn"},
-                     "IGNORE": {"fun": "load_from", "terms": "CompoundLensFilter/ThresholdEnergy"},
-                     "IGNORE": {"fun": "load_from", "terms": "PrivateFei/BitShift"},
-                     "IGNORE": {"fun": "load_from", "terms": "PrivateFei/DataBarSelected"},
-                     "IGNORE": {"fun": "load_from", "terms": "PrivateFei/DataBarAvailable"},
-                     "IGNORE": {"fun": "load_from", "terms": "PrivateFei/TimeOfCreation"},
-                     "IGNORE": {"fun": "load_from", "terms": "PrivateFei/DatabarHeight"},
-                     "IGNORE": {"fun": "load_from", "terms": "HiResIllumination/BrightFieldIsOn"},
-                     "IGNORE": {"fun": "load_from", "terms": "HiResIllumination/BrightFieldValue"},
-                     "IGNORE": {"fun": "load_from", "terms": "HiResIllumination/DarkFieldIsOn"},
-                     "IGNORE": {"fun": "load_from", "terms": "HiResIllumination/DarkFieldValue"},
-                     "IGNORE": {"fun": "load_from", "terms": "EasyLift/Rotation"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageMEMS/HeatingCurrent"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageMEMS/HeatingVoltage"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageMEMS/TargetTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageMEMS/ActualTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageMEMS/HeatingPower"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageMEMS/SampleBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageMEMS/SampleResistance"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStage/TargetTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStage/ActualTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStage/SampleBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStage/ShieldBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageHVHS/TargetTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageHVHS/ActualTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageHVHS/SampleBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "HotStageHVHS/ShieldBias"},
-                     "IGNORE": {"fun": "load_from", "terms": "ColdStage/TargetTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "ColdStage/ActualTemperature"},
-                     "IGNORE": {"fun": "load_from", "terms": "ColdStage/Humidity"},
-                     "IGNORE": {"fun": "load_from", "terms": "ColdStage/SampleBias"}}
+TIFF_TFS_TO_NEXUS_CFG = [('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/DETECTOR[detector*]/mode', 'load_from', 'Detectors/Mode'),
+                         ('/ENTRY[entry*]/measurement/em_lab/DETECTOR[detector*]/local_name', 'load_from', 'Detectors/Name'),
+                         ('/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/APERTURE_EM[aperture_em*]/description', 'load_from', 'EBeam/Aperture'),
+                         ('/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/APERTURE_EM[aperture_em*]/value', 'load_from', 'EBeam/ApertureDiameter'),
+                         ('/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/APERTURE_EM[aperture_em*]/value/@units', 'm'),
+                         ('/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/beam_current', 'load_from', 'EBeam/BeamCurrent'),
+                         ('/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/beam_current/@units', 'A'),
+                         ('/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/electron_source/voltage', 'load_from', 'EBeam/HV'),
+                         ('/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/electron_source/voltage/@units', 'V'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/STAGE_LAB[stage_lab]/tilt_1', 'load_from_rad_to_deg', 'EBeam/StageTa'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/STAGE_LAB[stage_lab]/tilt_1/@units', 'deg'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/STAGE_LAB[stage_lab]/tilt_2', 'load_from_rad_to_deg', 'EBeam/StageTb'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/STAGE_LAB[stage_lab]/tilt_2/@units', 'deg'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/EBEAM_COLUMN[ebeam_column]/operation_mode', 'load_from', 'EBeam/UseCase'),
+                         ('/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/working_distance', 'load_from', 'EBeam/WD'),
+                         ('/ENTRY[entry*]/measurement/em_lab/OPTICAL_SYSTEM_EM[optical_system_em]/working_distance/@units', 'm'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/event_type', 'load_from_lower_case', 'ETD/Signal'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/SCANBOX_EM[scanbox_em]/dwell_time', 'load_from', 'Scan/Dwelltime'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/SCANBOX_EM[scanbox_em]/dwell_time/@units', 's'),
+                         ('/ENTRY[entry*]/measurement/em_lab/FABRICATION[fabrication]/identifier', 'load_from', 'System/BuildNr'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/em_lab/SCANBOX_EM[scanbox_em]/scan_schema', 'load_from', 'System/Scan'),
+                         ('/ENTRY[entry*]/measurement/em_lab/EBEAM_COLUMN[ebeam_column]/electron_source/emitter_type', 'load_from', 'System/Source'),
+                         ('/ENTRY[entry*]/measurement/em_lab/FABRICATION[fabrication]/vendor', 'FEI'),
+                         ('/ENTRY[entry*]/measurement/em_lab/FABRICATION[fabrication]/model', 'load_from', 'System/SystemType'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/event_type', 'load_from_lower_case', 'T1/Signal'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/event_type', 'load_from_lower_case', 'T2/Signal'),
+                         ('/ENTRY[entry*]/measurement/EVENT_DATA_EM_SET[event_data_em_set]/EVENT_DATA_EM[event_data_em*]/event_type', 'load_from_lower_case', 'T3/Signal')]
