+++ pyproject.toml
@@ -28,6 +28,7 @@ dependencies = [
     "PyYAML>=6.0",
     "numpy>=1.21.2",
     "pandas>=1.3.2",
+    "odfpy>=1.4.1",
     "ase>=3.19.0",
     "flatdict>=4.0.1",
     "hyperspy>=1.7.6",
